Setup Vercel/Neon local database, removing the specific adapter
