+++ .gitignore
@@ -1,6 +1,8 @@
 node_modules
 .DS_Store
 .env
+.env.development
+.env.production
 
 # local PostgreSQL database
 db-data
