+++ docker-compose.yaml
@@ -1,41 +1,21 @@
-# A Docker Compose must always start with the version tag.
-# We use '3' because it's the last version.
-version: '3'
-
-# You should know that Docker Compose works with services.
-# 1 service = 1 container.
-# For example, a service, a server, a client, a database...
-# We use the keyword 'services' to start to create services.
+# A local Postgres instance coupled with a WebSocket proxy to mimic the serverless nature of Vercel Postgres
+# https://vercel.com/docs/storage/vercel-postgres/local-development
 services:
-  # The name of our service is "database"
-  # but you can use the name of your choice.
-  # Note: This may change the commands you are going to use a little bit.
-  database:
-    # Official Postgres image from DockerHub (we use the last version)
-    image: 'postgres:latest'
-
-    # By default, a Postgres database is running on the 5432 port.
-    # If we want to access the database from our computer (outside the container),
-    # we must share the port with our computer's port.
-    # The syntax is [port we want on our machine]:[port we want to retrieve in the container]
-    # Note: You are free to change your computer's port,
-    # but take into consideration that it will change the way
-    # you are connecting to your database.
+  postgres:
+    image: postgres
+    environment:
+      POSTGRES_PASSWORD: password
+      POSTGRES_DB: bestofjs-dev # The PostgreSQL default database (automatically created at first launch)
+    # Expose the Postgres port to the host machine,
+    # so you can inspect and administrate it
     ports:
-      - 5432:5432
-    # The `volumes` tag allows us to share a folder with our container
-    # Its syntax is as follows: [folder path on our machine]:[folder path to retrieve in the container]
+      - "54320:5432"
     volumes:
       # In this example, we share the folder `db-data` in our root repository, with the default PostgreSQL data path
       # It means that every time the repository is modifying the data inside
       # of `/var/lib/postgresql/data/`, automatically the change will appear in `db-data`
       # You don't need to create the `db-data` folder. Docker Compose will do it for you
-      - ./db-data/:/var/lib/postgresql/data/      
-
-    environment:
-      POSTGRES_USER: username # The PostgreSQL user (useful to connect to the database)
-      POSTGRES_PASSWORD: password # The PostgreSQL password (useful to connect to the database)
-      POSTGRES_DB: default_database # The PostgreSQL default database (automatically created at first launch)
+      - ./db-data/:/var/lib/postgresql/data/
   pg_proxy:
     image: ghcr.io/neondatabase/wsproxy:latest
     environment:
@@ -43,6 +23,8 @@ services:
       ALLOW_ADDR_REGEX: ".*"
       LOG_TRAFFIC: "true"
     ports:
-      - "5433:80"
+      # Expose the WebSocket proxy port to the host machine,
+      # this is where @vercel/postgres will connect
+      - "54330:80"
     depends_on:
-      - database      
+      - postgres
