+++ packages/db/package.json
@@ -3,9 +3,9 @@
   "private": true,
   "type": "module",
   "scripts": {
-    "generate": "dotenv -e ../../.env pnpm drizzle-kit generate",
-    "migrate": "dotenv -e ../../.env bun run src/migrate.ts",
-    "studio": "dotenv -e ../../.env pnpm drizzle-kit studio",
+    "generate": "dotenv -e ../../.env.${STAGE:=development} pnpm drizzle-kit generate",
+    "migrate": "bun run src/migrate.ts",
+    "studio": "dotenv -e ../../.env.${STAGE:=development} pnpm drizzle-kit studio",
     "typecheck": "tsc --noEmit",
     "typecheck:watch": "tsc --noEmit --incremental --watch",
     "format:write": "prettier --write \"src/**/*.{ts,tsx,mdx}\" --cache",
@@ -22,22 +22,22 @@
   },
   "dependencies": {
     "@clack/prompts": "^0.7.0",
+    "@neondatabase/serverless": "^0.9.5",
     "@t3-oss/env-nextjs": "^0.7.1",
     "@types/debug": "^4.1.12",
     "@types/lodash": "^4.14.195",
     "@types/luxon": "^3.4.2",
-    "@vercel/postgres": "^0.9.0",
+    "@vercel/postgres": "^0.10.0",
     "debug": "^4.3.4",
     "dotenv-cli": "^7.4.2",
-    "drizzle-kit": "^0.22.7",
-    "drizzle-orm": "^0.31.2",
+    "drizzle-kit": "^0.24.2",
+    "drizzle-orm": "^0.33.0",
     "emoji-regex": "^10.3.0",
     "graphql-request": "^7.0.1",
     "is-absolute-url": "^4.0.1",
     "lodash": "^4.17.21",
     "luxon": "^3.4.4",
     "nanoid": "^5.0.5",
-    "postgres": "^3.4.3",
     "scrape-it": "^6.1.2",
     "slugify": "^1.6.6",
     "tiny-invariant": "^1.3.1",
