--- packages/db/src/database.ts
@@ -1,4 +1,0 @@
-export interface DatabaseService<T> {
-  db: T;
-  disconnect(): void;
-}
